requirements.txt
Django==2.0.4
djangorestframework==3.8.2
freezegun==0.3.10
